Extract PhotosInputDataProvider to separate file
