+++ ChattoAdditions/Source/Input/Photos/PhotosInputWithPlaceholdersDataProvider.swift
@@ -28,7 +28,7 @@ protocol PhotosInputDataProviderDelegate: class {
     func handlePhotosInpudDataProviderUpdate(_ dataProvider: PhotosInputDataProviderProtocol, updateBlock: @escaping () -> Void)
 }
 
-protocol PhotosInputDataProviderProtocol : class {
+protocol PhotosInputDataProviderProtocol: class {
     weak var delegate: PhotosInputDataProviderDelegate? { get set }
     var count: Int { get }
     func requestPreviewImageAtIndex(_ index: Int, targetSize: CGSize, completion: @escaping (UIImage) -> Void) -> Int32
@@ -36,80 +36,7 @@ protocol PhotosInputDataProviderProtocol : class {
     func cancelPreviewImageRequest(_ requestID: Int32)
 }
 
-@objc
-class PhotosInputDataProvider: NSObject, PhotosInputDataProviderProtocol, PHPhotoLibraryChangeObserver {
-    weak var delegate: PhotosInputDataProviderDelegate?
-    private var imageManager = PHCachingImageManager()
-    private var fetchResult: PHFetchResult<PHAsset>!
-    override init() {
-        func fetchOptions(_ predicate: NSPredicate?) -> PHFetchOptions {
-            let options = PHFetchOptions()
-            options.sortDescriptors = [ NSSortDescriptor(key: "creationDate", ascending: false) ]
-            options.predicate = predicate
-            return options
-        }
-
-        if let userLibraryCollection = PHAssetCollection.fetchAssetCollections(with: .smartAlbum, subtype: .smartAlbumUserLibrary, options: nil).firstObject {
-            self.fetchResult = PHAsset.fetchAssets(in: userLibraryCollection, options: fetchOptions(NSPredicate(format: "mediaType = \(PHAssetMediaType.image.rawValue)")))
-        } else {
-            self.fetchResult = PHAsset.fetchAssets(with: .image, options: fetchOptions(nil))
-        }
-        super.init()
-        PHPhotoLibrary.shared().register(self)
-    }
-
-    deinit {
-        PHPhotoLibrary.shared().unregisterChangeObserver(self)
-    }
-
-    var count: Int {
-        return self.fetchResult.count
-    }
-
-    func requestPreviewImageAtIndex(_ index: Int, targetSize: CGSize, completion: @escaping (UIImage) -> Void) -> Int32 {
-        assert(index >= 0 && index < self.fetchResult.count, "Index out of bounds")
-        let asset = self.fetchResult[index]
-        let options = PHImageRequestOptions()
-        options.deliveryMode = .highQualityFormat
-        return self.imageManager.requestImage(for: asset, targetSize: targetSize, contentMode: .aspectFill, options: options) { (image, _) in
-            if let image = image {
-                completion(image)
-            }
-        }
-    }
-
-    func cancelPreviewImageRequest(_ requestID: Int32) {
-        self.imageManager.cancelImageRequest(requestID)
-    }
-
-    func requestFullImageAtIndex(_ index: Int, completion: @escaping (UIImage) -> Void) {
-        assert(index >= 0 && index < self.fetchResult.count, "Index out of bounds")
-        let asset = self.fetchResult[index]
-        self.imageManager.requestImageData(for: asset, options: .none) { (data, _, _, _) -> Void in
-            if let data = data, let image = UIImage(data: data) {
-                completion(image)
-            }
-        }
-    }
-
-    // MARK: PHPhotoLibraryChangeObserver
-
-    func photoLibraryDidChange(_ changeInstance: PHChange) {
-        // Photos may call this method on a background queue; switch to the main queue to update the UI.
-        DispatchQueue.main.async { [weak self]  in
-            guard let sSelf = self else { return }
-
-            if let changeDetails = changeInstance.changeDetails(for: sSelf.fetchResult as! PHFetchResult<PHObject>) {
-                let updateBlock = { () -> Void in
-                    self?.fetchResult = changeDetails.fetchResultAfterChanges as! PHFetchResult<PHAsset>
-                }
-                sSelf.delegate?.handlePhotosInpudDataProviderUpdate(sSelf, updateBlock: updateBlock)
-            }
-        }
-    }
-}
-
-class PhotosInputWithPlaceholdersDataProvider: PhotosInputDataProviderProtocol, PhotosInputDataProviderDelegate {
+final class PhotosInputWithPlaceholdersDataProvider: PhotosInputDataProviderProtocol, PhotosInputDataProviderDelegate {
     weak var delegate: PhotosInputDataProviderDelegate?
     private let photosDataProvider: PhotosInputDataProviderProtocol
     private let placeholdersDataProvider: PhotosInputDataProviderProtocol
