+++ ChattoAdditions/Source/Input/Photos/Placeholder/PhotosInputPlaceholderDataProvider.swift
@@ -26,24 +26,24 @@ import UIKit
 
 final class PhotosInputPlaceholderDataProvider: PhotosInputDataProviderProtocol {
     weak var delegate: PhotosInputDataProviderDelegate?
-    
+
     let numberOfPlaceholders: Int
-    
+
     init(numberOfPlaceholders: Int = 5) {
         self.numberOfPlaceholders = numberOfPlaceholders
     }
-    
+
     var count: Int {
         return self.numberOfPlaceholders
     }
-    
+
     func requestPreviewImageAtIndex(_ index: Int, targetSize: CGSize, completion: @escaping (UIImage) -> Void) -> Int32 {
         return 0
     }
-    
+
     func requestFullImageAtIndex(_ index: Int, completion: @escaping (UIImage) -> Void) {
     }
-    
+
     func cancelPreviewImageRequest(_ requestID: Int32) {
     }
 }
